Get Average Value of Statistical Life (VSL) for a Country

This function calculates the average Value of Statistical Life (VSL) for a specified country. It computes the weighted mean of VSL using the demography data as weights.

Usage

get_average_vsl(country)

Arguments

country

A string giving a country name from among daedalus::country_names or an ISO2 code from among daedalus::country_codes_iso2c or an ISO3 code from among daedalus::country_codes_iso3c.

Value

A numeric value representing the average VSL for the specified country.

Examples

if (FALSE) { # \dontrun{
  avg_vsl <- get_average_vsl("USA")
  print(avg_vsl)
} # }